modify cutting filter ranges getting as parameters

+++ src/iir.py
@@ -13,36 +13,36 @@ class FilterSignal:
             filtered_signal.append(float(filter[i]))
         return np.array(filtered_signal)
 
-    def butter_bandpass(self, lowcut: int, highcut: int):
+    def butter_bandpass(self, lowcut: float, highcut: float):
         nyq = 0.5 * self.fs
         low = lowcut / nyq
         high = highcut / nyq
         b, a = butter(self.order, [low, high], btype="band")
         return b, a
 
-    def butter_bandpass_filter(self, data, lowcut: int, highcut: int):
+    def butter_bandpass_filter(self, data, lowcut: float, highcut: float):
         b, a = self.butter_bandpass(lowcut, highcut)
         y = lfilter(b, a, data)
         return self.filtered_signal(y)
 
-    def butter_lowpass(self, cut: int):
+    def butter_lowpass(self, cut: float):
         nyq = 0.5 * self.fs
         ncut = cut / nyq
         b, a = butter(self.order, ncut, btype="low", analog=True)
         return b, a
 
-    def butter_lowpass_filter(self, data, cut: int):
+    def butter_lowpass_filter(self, data, cut: float):
         b, a = self.butter_lowpass(cut)
         y = lfilter(b, a, data)
         return self.filtered_signal(y)
 
-    def butter_highpass(self, cut: int):
+    def butter_highpass(self, cut: float):
         nyq = 0.5 * self.fs
         ncut = cut / nyq
         b, a = butter(self.order, ncut, btype="high", analog=True)
         return b, a
 
-    def butter_highpass_filter(self, data, cut: int):
+    def butter_highpass_filter(self, data, cut: float):
         b, a = self.butter_highpass(cut)
         y = lfilter(b, a, data)
         return self.filtered_signal(y)
